backend/requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
pillow==10.2.0
torch==2.1.2
transformers==4.36.2
huggingface-hub==0.20.3
matplotlib==3.8.2
aiofiles==23.2.1
pytest==7.4.3
httpx==0.26.0